+++ StarLevelSystem/common/DataObjects.cs
@@ -1,13 +1,10 @@
-﻿using Jotunn.Configs;
-using Jotunn.Entities;
+﻿using Jotunn.Entities;
 using Jotunn.Managers;
 using StarLevelSystem.Data;
-using StarLevelSystem.modules;
 using System;
 using System.Collections.Generic;
 using System.ComponentModel;
 using System.IO;
-using System.Linq;
 using System.Reflection;
 using System.Runtime.Serialization;
 using System.Runtime.Serialization.Formatters.Binary;